+++ tests/_config/bootstrap.php
@@ -47,6 +47,7 @@ $loader = new Loader();
 $loader->registerNamespaces(
     [
         'Phalcon\Test\Models'      => $config->get('application')->modelsDir,
+        'Phalcon\Test\Resultsets'  => $config->get('application')->resultsetsDir,
         'Phalcon\Test\Collections' => $config->get('application')->collectionsDir,
         'Phalcon\Test\Modules\Frontend\Controllers' => $config->get('application')->modulesDir . 'frontend/controllers/',
         'Phalcon\Test\Modules\Backend\Controllers'  => $config->get('application')->modulesDir . 'backend/controllers/'